LPS-102342 Remove unnecessary explicit output
All of these projects use the liferay-npm-bundler-preset-liferay-dev
preset because they use liferay-npm-scripts, and that means they get the
correct "output" setting by default.

The exception is the dynamic-data-mapping-form-taglib project that seems
to have an .npmbundlerrc file via erroneous copy-paste (there is no JS
in this project at all, nor a package.json).